refactor(broken-link-checker): read checker options from a config file
