+++ broken-link-checker.config.js
@@ -1,0 +1,1 @@
+module.exports = {};
