+++ packages/broken-link-checker/package.json
@@ -13,10 +13,11 @@
     "url": "https://github.com/commercetools/commercetools-docs-kit.git"
   },
   "bin": {
-    "commercetools-broken-link-checker": "broken-link-checker.js"
+    "commercetools-broken-link-checker": "bin/broken-link-checker.js"
   },
   "dependencies": {
     "broken-link-checker": "0.7.8",
+    "cosmiconfig": "6.0.0",
     "express": "4.17.1",
     "get-port": "5.0.0",
     "mri": "1.1.4"
